Time: 4 ms (83.29%), Space: 7.9 MB (90.03%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.c
@@ -2,11 +2,11 @@ bool isPalindrome(int x) {
     if(x<0){
         return false;
     }
-    unsigned int j = 1;
-    while(j <= (unsigned int)x/10){
+    int j = 1;
+    while(j <= x/10){
         j *= 10;
     }
-    unsigned int i = 1;
+    int i = 1;
     while(j>=i){
         if((x%(i*10))/i != (x/j)%10){
             return false;
